Added user data property to stream_ctx
Also moved stream_key into its own file to follow the same pattern of other objects
