+++ docs/html/files.html
@@ -86,12 +86,13 @@ $(function() {
 <tr id="row_0_0_17_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="protocol_8h_source.html"><span class="icondoc"></span></a><b>protocol.h</b></td><td class="desc"></td></tr>
 <tr id="row_0_0_18_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="root__keys_8h_source.html"><span class="icondoc"></span></a><b>root_keys.h</b></td><td class="desc"></td></tr>
 <tr id="row_0_0_19_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="storage_8h_source.html"><span class="icondoc"></span></a><b>storage.h</b></td><td class="desc"></td></tr>
-<tr id="row_0_0_20_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="stream__cipher_8h_source.html"><span class="icondoc"></span></a><b>stream_cipher.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_20_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="stream__ctx_8h_source.html"><span class="icondoc"></span></a><b>stream_ctx.h</b></td><td class="desc"></td></tr>
 <tr id="row_0_0_21_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="stream__iv_8h_source.html"><span class="icondoc"></span></a><b>stream_iv.h</b></td><td class="desc"></td></tr>
-<tr id="row_0_0_22_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="transport__ctx_8h_source.html"><span class="icondoc"></span></a><b>transport_ctx.h</b></td><td class="desc"></td></tr>
-<tr id="row_0_0_23_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="util_8h_source.html"><span class="icondoc"></span></a><b>util.h</b></td><td class="desc"></td></tr>
-<tr id="row_0_0_24_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="wickr-crypto-c_8h_source.html"><span class="icondoc"></span></a><b>wickr-crypto-c.h</b></td><td class="desc"></td></tr>
-<tr id="row_0_0_25_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="wickr__ctx_8h_source.html"><span class="icondoc"></span></a><b>wickr_ctx.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_22_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="stream__key_8h_source.html"><span class="icondoc"></span></a><b>stream_key.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_23_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="transport__ctx_8h_source.html"><span class="icondoc"></span></a><b>transport_ctx.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_24_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="util_8h_source.html"><span class="icondoc"></span></a><b>util.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_25_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="wickr-crypto-c_8h_source.html"><span class="icondoc"></span></a><b>wickr-crypto-c.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_26_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="wickr__ctx_8h_source.html"><span class="icondoc"></span></a><b>wickr_ctx.h</b></td><td class="desc"></td></tr>
 </table>
 </div><!-- directory -->
 </div><!-- contents -->
