+++ docs/html/functions_vars.html
@@ -441,6 +441,13 @@ $(function() {
 </ul>
 
 
+<h3><a id="index_u"></a>- u -</h3><ul>
+<li>user_data
+: <a class="el" href="structwickr__stream__key.html#a51e59da9f7fbb0b3e1332a20252d1cc6">wickr_stream_key</a>
+</li>
+</ul>
+
+
 <h3><a id="index_v"></a>- v -</h3><ul>
 <li>version
 : <a class="el" href="structwickr__packet.html#a35c17c926c267a603e810b88961c2bf3">wickr_packet</a>
