+++ docs/html/modules.html
@@ -86,9 +86,10 @@ $(function() {
 <tr id="row_19_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__storage__keys.html" target="_self">Wickr_storage_keys</a></td><td class="desc"></td></tr>
 <tr id="row_20_" class="even"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__stream.html" target="_self">wickr_stream</a></td><td class="desc"></td></tr>
 <tr id="row_21_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__stream__iv.html" target="_self">Wickr_stream_iv</a></td><td class="desc"></td></tr>
-<tr id="row_22_" class="even"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__transport__ctx.html" target="_self">Wickr_transport_ctx</a></td><td class="desc"></td></tr>
-<tr id="row_23_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__data__conversion__utilities.html" target="_self">Data Conversion Utilities</a></td><td class="desc"></td></tr>
-<tr id="row_24_" class="even"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__ctx.html" target="_self">Wickr_ctx</a></td><td class="desc"></td></tr>
+<tr id="row_22_" class="even"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__stream__key.html" target="_self">wickr_stream_key</a></td><td class="desc"></td></tr>
+<tr id="row_23_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__transport__ctx.html" target="_self">Wickr_transport_ctx</a></td><td class="desc"></td></tr>
+<tr id="row_24_" class="even"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__data__conversion__utilities.html" target="_self">Data Conversion Utilities</a></td><td class="desc"></td></tr>
+<tr id="row_25_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><a class="el" href="group__wickr__ctx.html" target="_self">Wickr_ctx</a></td><td class="desc"></td></tr>
 </table>
 </div><!-- directory -->
 </div><!-- contents -->
