+++ docs/html/search/all_13.js
@@ -1,268 +1,4 @@
 var searchData=
 [
-  ['wickr_5falloc',['wickr_alloc',['../group__memory__functions.html#ga227aa24a951a89daf9058c30ef525836',1,'memory.h']]],
-  ['wickr_5falloc_5fzero',['wickr_alloc_zero',['../group__memory__functions.html#ga8163f2ae01a21e319806fbba9b9da00a',1,'memory.h']]],
-  ['wickr_5farray_5ft',['wickr_array_t',['../group__wickr__array.html',1,'']]],
-  ['wickr_5farray_5fcopy',['wickr_array_copy',['../group__wickr__array.html#ga1e28a815bd4310985e493b67f66f26c7',1,'array.h']]],
-  ['wickr_5farray_5fdestroy',['wickr_array_destroy',['../group__wickr__array.html#ga984338e0a278ef1b23aca7330f0bfcc3',1,'array.h']]],
-  ['wickr_5farray_5ffetch_5fitem',['wickr_array_fetch_item',['../group__wickr__array.html#ga3a6dd3ee6295f6230e2047584a261774',1,'array.h']]],
-  ['wickr_5farray_5fget_5fitem_5fcount',['wickr_array_get_item_count',['../group__wickr__array.html#gafe1191ce494529dbfd5240a84bcba5a2',1,'array.h']]],
-  ['wickr_5farray_5fnew',['wickr_array_new',['../group__wickr__array.html#gad5842105c20b42cc4b5b4ffc2a72556b',1,'array.h']]],
-  ['wickr_5farray_5fset_5fitem',['wickr_array_set_item',['../group__wickr__array.html#ga3b3c15b3c6a15f52d20c678b6829ee21',1,'array.h']]],
-  ['wickr_5fbuffer',['wickr_buffer',['../structwickr__buffer.html',1,'wickr_buffer'],['../group__wickr__buffer.html',1,'(Global Namespace)']]],
-  ['wickr_5fbuffer_5fconcat',['wickr_buffer_concat',['../group__wickr__buffer.html#gad7721c1033a1fc378a0a11613f3cfb26',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fconcat_5fmulti',['wickr_buffer_concat_multi',['../group__wickr__buffer.html#ga3aa7ea751a65310184d6c0ea9627dca8',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fcopy',['wickr_buffer_copy',['../group__wickr__buffer.html#gada179dda91e748d7bdf6028d3d4c4bcd',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fcopy_5fsection',['wickr_buffer_copy_section',['../group__wickr__buffer.html#gaceb6345c35ac2f6330ea0a685ce3fc53',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fcreate',['wickr_buffer_create',['../group__wickr__buffer.html#gacc3ad1220af28781bf678cab20e2f1c8',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fcreate_5fempty',['wickr_buffer_create_empty',['../group__wickr__buffer.html#ga305edd191e38eb040710b631c1e4f05f',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fcreate_5fempty_5fzero',['wickr_buffer_create_empty_zero',['../group__wickr__buffer.html#gaba0ae148a2f75e50eb3d36a4bef509d2',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fdestroy',['wickr_buffer_destroy',['../group__wickr__buffer.html#ga7ec1198db0739f3034ddd3969c64257f',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fdestroy_5fzero',['wickr_buffer_destroy_zero',['../group__wickr__buffer.html#gaf0ff6486debbfd676b7675f6f899fc40',1,'buffer.h']]],
-  ['wickr_5fbuffer_5ffrom_5fprotobytes',['wickr_buffer_from_protobytes',['../group__protobuf__utils.html#ga82aef3139be67d3646dc1e76c31a3352',1,'protobuf_util.h']]],
-  ['wickr_5fbuffer_5fis_5fequal',['wickr_buffer_is_equal',['../group__wickr__buffer.html#ga51ec58848ab0b07d164c6f7d73200c24',1,'buffer.h']]],
-  ['wickr_5fbuffer_5fmodify_5fsection',['wickr_buffer_modify_section',['../group__wickr__buffer.html#ga67f817bf762e4ee401b31685197620be',1,'buffer.h']]],
-  ['wickr_5fcipher',['wickr_cipher',['../structwickr__cipher.html',1,'wickr_cipher'],['../group__wickr__cipher.html',1,'(Global Namespace)']]],
-  ['wickr_5fcipher_5ffind',['wickr_cipher_find',['../group__wickr__cipher.html#ga2b7a8bc33e1fd5af61763398ea9a4013',1,'cipher.h']]],
-  ['wickr_5fcipher_5fkey',['wickr_cipher_key',['../structwickr__cipher__key.html',1,'']]],
-  ['wickr_5fcipher_5fkey_5fcopy',['wickr_cipher_key_copy',['../group__wickr__cipher.html#gaa473bc83f82d6353aa6f4cb4e36bf4f0',1,'cipher.h']]],
-  ['wickr_5fcipher_5fkey_5fcreate',['wickr_cipher_key_create',['../group__wickr__cipher.html#ga73e2a916ba0ff19517e30a65a68390f0',1,'cipher.h']]],
-  ['wickr_5fcipher_5fkey_5fdestroy',['wickr_cipher_key_destroy',['../group__wickr__cipher.html#ga45790b3023ccb261c03346b069aca763',1,'cipher.h']]],
-  ['wickr_5fcipher_5fkey_5ffrom_5fbuffer',['wickr_cipher_key_from_buffer',['../group__wickr__cipher.html#ga1c073b87f2141cf35a1d91ba10f9e5ba',1,'cipher.h']]],
-  ['wickr_5fcipher_5fkey_5ffrom_5fprotobytes',['wickr_cipher_key_from_protobytes',['../group__protobuf__utils.html#ga29c41b9d38441ba0fc54c7ffc0ea91a4',1,'protobuf_util.h']]],
-  ['wickr_5fcipher_5fkey_5fserialize',['wickr_cipher_key_serialize',['../group__wickr__cipher.html#ga8716aebf03497c379d5ff81fe32cde32',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult',['wickr_cipher_result',['../structwickr__cipher__result.html',1,'']]],
-  ['wickr_5fcipher_5fresult_5fcopy',['wickr_cipher_result_copy',['../group__wickr__cipher.html#gace82050fcc8b7932df9a9e8109a2fc1d',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult_5fcreate',['wickr_cipher_result_create',['../group__wickr__cipher.html#ga234b60967d0b3ef0088ca3ac30755dde',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult_5fdestroy',['wickr_cipher_result_destroy',['../group__wickr__cipher.html#gad5d22ca53abc4397d88040f0f8003021',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult_5ffrom_5fbuffer',['wickr_cipher_result_from_buffer',['../group__wickr__cipher.html#gad42314812c457855c3799632494c91a7',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult_5fis_5fvalid',['wickr_cipher_result_is_valid',['../group__wickr__cipher.html#ga3f1b679c199987148952816b6fc4c009',1,'cipher.h']]],
-  ['wickr_5fcipher_5fresult_5fserialize',['wickr_cipher_result_serialize',['../group__wickr__cipher.html#gacda386115dc52f83a0518f8ac0dad6a2',1,'cipher.h']]],
-  ['wickr_5fcrypto_5fengine',['wickr_crypto_engine',['../structwickr__crypto__engine.html',1,'wickr_crypto_engine'],['../group__wickr__crypto__engine.html',1,'(Global Namespace)']]],
-  ['wickr_5fcrypto_5fengine_5fcipher_5fdecrypt',['wickr_crypto_engine_cipher_decrypt',['../group__wickr__crypto__engine.html#ga459ef821cc40edb5e3c9ce1d571ddc00',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcipher_5fencrypt',['wickr_crypto_engine_cipher_encrypt',['../group__wickr__crypto__engine.html#gaf7be0d0ca0d9e7a0c5c3b1ec0fd0918e',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcipher_5fkey_5frandom',['wickr_crypto_engine_cipher_key_random',['../group__wickr__crypto__engine.html#gabc120effc5248a491b18138f9b74b526',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcrypto_5frandom',['wickr_crypto_engine_crypto_random',['../group__wickr__crypto__engine.html#gae12855853c05caa473e748a62611d307',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdecrypt_5ffile',['wickr_crypto_engine_decrypt_file',['../group__wickr__crypto__engine.html#ga361b843efc839e30f90ce24c15b1dd27',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdigest',['wickr_crypto_engine_digest',['../group__wickr__crypto__engine.html#ga42dfc5b4df166616d962ee4c792ceaf4',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdigest_5ffile',['wickr_crypto_engine_digest_file',['../group__wickr__crypto__engine.html#ga298d2dea6722151c11e04458d3924823',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fkey_5fimport',['wickr_crypto_engine_ec_key_import',['../group__wickr__crypto__engine.html#ga87a069bbd3cdf509946e4ca0ca008f8f',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5frand_5fkey',['wickr_crypto_engine_ec_rand_key',['../group__wickr__crypto__engine.html#gad16b229b7d16cf00d3983c277ee3b88b',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fsign',['wickr_crypto_engine_ec_sign',['../group__wickr__crypto__engine.html#gac5a36d2a53e06a75a2b818e6a5cc96a1',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fverify',['wickr_crypto_engine_ec_verify',['../group__wickr__crypto__engine.html#gae541acc6433f6f5ffe84382ccd5e4f97',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fecdh_5fgen_5fkey',['wickr_crypto_engine_ecdh_gen_key',['../group__wickr__crypto__engine.html#ga94bdc8d3b88ac04b9defe663afa48e78',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fencrypt_5ffile',['wickr_crypto_engine_encrypt_file',['../group__wickr__crypto__engine.html#gaa77b1bc373d92d9db9d801bcc4a4d29c',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fget_5fdefault',['wickr_crypto_engine_get_default',['../group__wickr__crypto__engine.html#ga74d5cea6c7d1d6e67b1be476fbaf6235',1,'crypto_engine.h']]],
-  ['wickr_5fcrypto_5fengine_5fhmac_5fcreate',['wickr_crypto_engine_hmac_create',['../group__wickr__crypto__engine.html#ga782e0e8786ecf8a55c8ae2aea2f178aa',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fhmac_5fverify',['wickr_crypto_engine_hmac_verify',['../group__wickr__crypto__engine.html#ga00c52a816403192af5b21d952265d0b6',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fkdf_5fcipher',['wickr_crypto_engine_kdf_cipher',['../group__wickr__crypto__engine.html#ga5709d94cd09ef0fb47d1a5f59232f115',1,'crypto_engine.h']]],
-  ['wickr_5fcrypto_5fengine_5fkdf_5fdecipher',['wickr_crypto_engine_kdf_decipher',['../group__wickr__crypto__engine.html#gab60480270e80ba34aa44385633037983',1,'crypto_engine.h']]],
-  ['wickr_5fcrypto_5fkdf_5fgen',['wickr_crypto_kdf_gen',['../group__wickr__crypto__engine.html#ga5ab29968c993423df83c9f1a3d7e685d',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fkdf_5fmeta',['wickr_crypto_kdf_meta',['../group__wickr__crypto__engine.html#ga640cbc86ca36c4deb2c197de89dce7d1',1,'wickr_crypto_engine']]],
-  ['wickr_5fctx',['wickr_ctx',['../structwickr__ctx.html',1,'wickr_ctx'],['../group__wickr__ctx.html',1,'(Global Namespace)']]],
-  ['wickr_5fctx_5fcipher_5flocal',['wickr_ctx_cipher_local',['../group__wickr__ctx.html#gaeff6704dded2d44f689eca5ad25843f8',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fcipher_5fremote',['wickr_ctx_cipher_remote',['../group__wickr__ctx.html#ga0b985fb2cd05fb482b5ed175dfd650a0',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fcopy',['wickr_ctx_copy',['../group__wickr__ctx.html#gaaa74a32619fb3b382fbeeeddc425a910',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fcreate',['wickr_ctx_create',['../group__wickr__ctx.html#ga085975d71519bf5729a7cf264a5f7876',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fdecipher_5fremote',['wickr_ctx_decipher_remote',['../group__wickr__ctx.html#gacc1038d3690ba4860b74f13239c65600',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fdestroy',['wickr_ctx_destroy',['../group__wickr__ctx.html#gaadfdef183c86992efc33b0d730c90327',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fencode',['wickr_ctx_encode',['../structwickr__ctx__encode.html',1,'']]],
-  ['wickr_5fctx_5fencode_5fcreate',['wickr_ctx_encode_create',['../group__wickr__ctx.html#ga810ee459cb0ab1037b5fd1f983d0ded3',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fencode_5fdestroy',['wickr_ctx_encode_destroy',['../group__wickr__ctx.html#ga9ed3f7c0f4d672a2ac562b279e2da4e1',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fencode_5fpacket',['wickr_ctx_encode_packet',['../group__wickr__ctx.html#ga3766720fd6654fad791a26cd2c1161b7',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fephemeral_5fkeypair_5fgen',['wickr_ctx_ephemeral_keypair_gen',['../group__wickr__ctx.html#ga45b2531085c430942122dcbf45736e10',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fexport_5fstorage_5fkeys',['wickr_ctx_export_storage_keys',['../group__wickr__ctx.html#ga67e64bed3fc0a8ee550c0fc24d44c787',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fexport_5frecovery_5fkey_5fpassphrase',['wickr_ctx_gen_export_recovery_key_passphrase',['../group__wickr__ctx.html#ga2dadd40adc0217dd337ee6cf745cbc12',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fimport_5frecovery',['wickr_ctx_gen_import_recovery',['../group__wickr__ctx.html#ga0bde95494141673604b9e4e1f7954e2c',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fnew',['wickr_ctx_gen_new',['../group__wickr__ctx.html#gaa8904ab3687a22f8d291f5976d05a0e5',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fnew_5fwith_5fsig_5fkey',['wickr_ctx_gen_new_with_sig_key',['../group__wickr__ctx.html#ga458f74fb060fb98da491d733e51fc3ef',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fresult',['wickr_ctx_gen_result',['../structwickr__ctx__gen__result.html',1,'']]],
-  ['wickr_5fctx_5fgen_5fresult_5fcopy',['wickr_ctx_gen_result_copy',['../group__wickr__ctx.html#gae095a8cb40d209090212136f8580dc77',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fresult_5fdestroy',['wickr_ctx_gen_result_destroy',['../group__wickr__ctx.html#ga865854408d22e4697eebe5bc129cc6b1',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fresult_5fmake_5frecovery',['wickr_ctx_gen_result_make_recovery',['../group__wickr__ctx.html#gaa32727731391f1e3fbde26094924b533',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fwith_5fpassphrase',['wickr_ctx_gen_with_passphrase',['../group__wickr__ctx.html#gaab0b2731d76d6ea91afd1b147c53a29a',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fwith_5frecovery',['wickr_ctx_gen_with_recovery',['../group__wickr__ctx.html#ga2e9c63af0e76f1977844ea2dc895e90a',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fgen_5fwith_5froot_5fkeys',['wickr_ctx_gen_with_root_keys',['../group__wickr__ctx.html#ga286a513a606b6643f796539413357c34',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fimport_5fstorage_5fkeys',['wickr_ctx_import_storage_keys',['../group__wickr__ctx.html#gae344e5202193c4b92ee2686529285fe5',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fpacket',['wickr_ctx_packet',['../structwickr__ctx__packet.html',1,'']]],
-  ['wickr_5fctx_5fpacket_5fcreate',['wickr_ctx_packet_create',['../group__wickr__ctx.html#ga8ac1e1c9a0b9802d7f259278a09df56c',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fpacket_5fdestroy',['wickr_ctx_packet_destroy',['../group__wickr__ctx.html#gad7063bcfca4878d290e598ed6bc22e3c',1,'wickr_ctx.h']]],
-  ['wickr_5fctx_5fparse_5fpacket',['wickr_ctx_parse_packet',['../group__wickr__ctx.html#ga5032fb65627a509fd9486ea9696e643d',1,'wickr_ctx.h']]],
-  ['wickr_5fdecode_5fresult',['wickr_decode_result',['../structwickr__decode__result.html',1,'']]],
-  ['wickr_5fdecode_5fresult_5fcopy',['wickr_decode_result_copy',['../group__wickr__protocol.html#ga36183bf96372661f890aefbe388bd32a',1,'protocol.h']]],
-  ['wickr_5fdecode_5fresult_5fcreate_5ffailure',['wickr_decode_result_create_failure',['../group__wickr__protocol.html#ga2bd6f49ed993d2290aaafa514eba43ed',1,'protocol.h']]],
-  ['wickr_5fdecode_5fresult_5fcreate_5fsuccess',['wickr_decode_result_create_success',['../group__wickr__protocol.html#ga633aaf4cb6159be0110401f06dae458c',1,'protocol.h']]],
-  ['wickr_5fdecode_5fresult_5fdestroy',['wickr_decode_result_destroy',['../group__wickr__protocol.html#ga94154ab44a6c2e9b97a115e37416efe1',1,'protocol.h']]],
-  ['wickr_5fdecode_5fresult_5ffrom_5fparse_5fresult',['wickr_decode_result_from_parse_result',['../group__wickr__protocol.html#ga32d9da2e10d1d653b2ac894d50aaedb0',1,'protocol.h']]],
-  ['wickr_5fdev_5finfo',['wickr_dev_info',['../structwickr__dev__info.html',1,'wickr_dev_info'],['../group__wickr__dev__info.html',1,'(Global Namespace)']]],
-  ['wickr_5fdev_5finfo_5fcopy',['wickr_dev_info_copy',['../group__wickr__dev__info.html#ga1c9d2bdc1778166c49aec43906bb1a83',1,'devinfo.h']]],
-  ['wickr_5fdev_5finfo_5fcreate',['wickr_dev_info_create',['../group__wickr__dev__info.html#gaadd5ceb4c6ed757440332e8e27109e5e',1,'devinfo.h']]],
-  ['wickr_5fdev_5finfo_5fcreate_5fnew',['wickr_dev_info_create_new',['../group__wickr__dev__info.html#gadbe4731cffb188da317ed9237826ac5d',1,'devinfo.h']]],
-  ['wickr_5fdev_5finfo_5fderive',['wickr_dev_info_derive',['../group__wickr__dev__info.html#ga7f81b96fc222560762d608772b3f44bd',1,'devinfo.h']]],
-  ['wickr_5fdev_5finfo_5fdestroy',['wickr_dev_info_destroy',['../group__wickr__dev__info.html#ga09ee324a0e0e6a0451aa53031eecb324',1,'devinfo.h']]],
-  ['wickr_5fdigest',['wickr_digest',['../structwickr__digest.html',1,'wickr_digest'],['../group__wickr__digest.html',1,'(Global Namespace)']]],
-  ['wickr_5fdigest_5ffind_5fwith_5fid',['wickr_digest_find_with_id',['../group__wickr__digest.html#ga64a18a57a861988bf1a4898bebc712cb',1,'digest.h']]],
-  ['wickr_5fdigest_5fmatching_5fcipher',['wickr_digest_matching_cipher',['../group__wickr__crypto__engine.html#ga18c21501a41c97f2ba79a9366343cdcf',1,'crypto_engine.h']]],
-  ['wickr_5fdigest_5fmatching_5fcurve',['wickr_digest_matching_curve',['../group__wickr__crypto__engine.html#ga0a708e5e46c40eae5eaeb581f6a1aed2',1,'crypto_engine.h']]],
-  ['wickr_5fec_5fcurve',['wickr_ec_curve',['../structwickr__ec__curve.html',1,'wickr_ec_curve'],['../group__wickr__ec__curve.html',1,'(Global Namespace)']]],
-  ['wickr_5fec_5fcurve_5ffind',['wickr_ec_curve_find',['../group__wickr__ec__curve.html#ga35d0968431b69993b7a6e3351d503d72',1,'eckey.h']]],
-  ['wickr_5fec_5fkey',['wickr_ec_key',['../structwickr__ec__key.html',1,'']]],
-  ['wickr_5fec_5fkey_5fcopy',['wickr_ec_key_copy',['../group__wickr__ec__curve.html#gaf12eff11d78b2ce9ff32710406d58d4f',1,'eckey.h']]],
-  ['wickr_5fec_5fkey_5fcreate',['wickr_ec_key_create',['../group__wickr__ec__curve.html#ga912f46bc12a39f57d8d1b87a639f7cdf',1,'eckey.h']]],
-  ['wickr_5fec_5fkey_5fdestroy',['wickr_ec_key_destroy',['../group__wickr__ec__curve.html#gac18ac86b270f540bf5a55ad271ab0568',1,'eckey.h']]],
-  ['wickr_5fec_5fkey_5ffrom_5fprotobytes',['wickr_ec_key_from_protobytes',['../group__protobuf__utils.html#gac18223ab7af7d533ac09084f3c2caa4b',1,'protobuf_util.h']]],
-  ['wickr_5fecdh_5fparams',['wickr_ecdh_params',['../structwickr__ecdh__params.html',1,'wickr_ecdh_params'],['../group__wickr__ecdh__params.html',1,'(Global Namespace)']]],
-  ['wickr_5fecdh_5fparams_5fare_5fvalid',['wickr_ecdh_params_are_valid',['../group__wickr__ecdh__params.html#ga801f1bfae11ac0c94d59bf9dd58f78aa',1,'ecdh.h']]],
-  ['wickr_5fecdh_5fparams_5fcopy',['wickr_ecdh_params_copy',['../group__wickr__ecdh__params.html#ga016f2b643f2636c05ed72bc08a0314f0',1,'ecdh.h']]],
-  ['wickr_5fecdh_5fparams_5fcreate',['wickr_ecdh_params_create',['../group__wickr__ecdh__params.html#ga8a5be72f13caa4bf5ee99e16cc3b092f',1,'ecdh.h']]],
-  ['wickr_5fecdh_5fparams_5fdestroy',['wickr_ecdh_params_destroy',['../group__wickr__ecdh__params.html#gaf6efcaf58f603e323443063cde9b5621',1,'ecdh.h']]],
-  ['wickr_5fecdsa_5fresult',['wickr_ecdsa_result',['../structwickr__ecdsa__result.html',1,'wickr_ecdsa_result'],['../group__wickr__ecdsa__result.html',1,'(Global Namespace)']]],
-  ['wickr_5fecdsa_5fresult_5fcopy',['wickr_ecdsa_result_copy',['../group__wickr__ecdsa__result.html#ga8664bbdec0bc8b354b3cb35704c42f57',1,'ecdsa.h']]],
-  ['wickr_5fecdsa_5fresult_5fcreate',['wickr_ecdsa_result_create',['../group__wickr__ecdsa__result.html#ga7e199a7c851fd07fb93180fb3a242749',1,'ecdsa.h']]],
-  ['wickr_5fecdsa_5fresult_5fcreate_5ffrom_5fbuffer',['wickr_ecdsa_result_create_from_buffer',['../group__wickr__ecdsa__result.html#gaa3a5ba5ce3df076cf5a9984f68e34c3a',1,'ecdsa.h']]],
-  ['wickr_5fecdsa_5fresult_5fdestroy',['wickr_ecdsa_result_destroy',['../group__wickr__ecdsa__result.html#ga84fb120b34b4b7ede4042009c735ea5d',1,'ecdsa.h']]],
-  ['wickr_5fecdsa_5fresult_5ffrom_5fprotobytes',['wickr_ecdsa_result_from_protobytes',['../group__protobuf__utils.html#ga3b5c6c2941820a16a527fccebb0b8821',1,'protobuf_util.h']]],
-  ['wickr_5fecdsa_5fresult_5fserialize',['wickr_ecdsa_result_serialize',['../group__wickr__ecdsa__result.html#ga417d1186e93d6ae4ee4cf701a2bfb1b2',1,'ecdsa.h']]],
-  ['wickr_5fephemeral_5finfo',['wickr_ephemeral_info',['../structwickr__ephemeral__info.html',1,'']]],
-  ['wickr_5fephemeral_5fkeypair',['wickr_ephemeral_keypair',['../structwickr__ephemeral__keypair.html',1,'wickr_ephemeral_keypair'],['../group__wickr__ephemeral__keypair.html',1,'(Global Namespace)']]],
-  ['wickr_5fephemeral_5fkeypair_5fcopy',['wickr_ephemeral_keypair_copy',['../group__wickr__ephemeral__keypair.html#gacf14f050221091093b9a2948bd569662',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fcreate',['wickr_ephemeral_keypair_create',['../group__wickr__ephemeral__keypair.html#gab956214203938b4b02b8f30666eaec3b',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fcreate_5ffrom_5fbuffer',['wickr_ephemeral_keypair_create_from_buffer',['../group__wickr__ephemeral__keypair.html#ga7ecd3769c68ba5bfbc95a225c1021ed0',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fdestroy',['wickr_ephemeral_keypair_destroy',['../group__wickr__ephemeral__keypair.html#gab77107c0b1a0d145c606817eeecec63d',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fgenerate_5fidentity',['wickr_ephemeral_keypair_generate_identity',['../group__wickr__ephemeral__keypair.html#ga6c0935c579e3d4bdbfeada7cb60c2a1b',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fmake_5fpublic',['wickr_ephemeral_keypair_make_public',['../group__wickr__ephemeral__keypair.html#ga2e2f8db184aa2972ef5423e9c03988e8',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fserialize',['wickr_ephemeral_keypair_serialize',['../group__wickr__ephemeral__keypair.html#ga1c994a08f270cc84088263d48a4b3636',1,'ephemeral_keypair.h']]],
-  ['wickr_5fephemeral_5fkeypair_5fverify_5fowner',['wickr_ephemeral_keypair_verify_owner',['../group__wickr__ephemeral__keypair.html#gadc17110d4c7fa65a9456fae516bf2e5d',1,'ephemeral_keypair.h']]],
-  ['wickr_5fexchange_5farray_5fcopy',['wickr_exchange_array_copy',['../group__wickr__protocol.html#gacd04405337871aa3bcd550a051d778e1',1,'protocol.h']]],
-  ['wickr_5fexchange_5farray_5fdestroy',['wickr_exchange_array_destroy',['../group__wickr__protocol.html#gadc1e5f5836bc09c1d6b8d6203d02a0eb',1,'protocol.h']]],
-  ['wickr_5fexchange_5farray_5ffetch_5fitem',['wickr_exchange_array_fetch_item',['../group__wickr__protocol.html#ga333510c3c7eeb8e33f94495de6bb122d',1,'protocol.h']]],
-  ['wickr_5fexchange_5farray_5fnew',['wickr_exchange_array_new',['../group__wickr__protocol.html#ga907d700b76b748f80796b4404c7172c2',1,'protocol.h']]],
-  ['wickr_5fexchange_5farray_5fset_5fitem',['wickr_exchange_array_set_item',['../group__wickr__protocol.html#gad3c6687c07ae5e66efcbc55da0febd85',1,'protocol.h']]],
-  ['wickr_5fexchange_5fcipher_5fmatching_5fcipher',['wickr_exchange_cipher_matching_cipher',['../group__wickr__crypto__engine.html#gaf44e3be9066727bf83034d3593dc74e2',1,'crypto_engine.h']]],
-  ['wickr_5ffree',['wickr_free',['../group__memory__functions.html#ga0ec3692f8dc5c2df128743c515d300c1',1,'memory.h']]],
-  ['wickr_5ffree_5fzero',['wickr_free_zero',['../group__memory__functions.html#gadad0d18d3a00f50082b6585971a62ae1',1,'memory.h']]],
-  ['wickr_5fhkdf_5falgo_5ffor_5fdigest',['wickr_hkdf_algo_for_digest',['../group__wickr__kdf.html#ga81c7a5adea69d04e019dd58080268a08',1,'kdf.h']]],
-  ['wickr_5fidentity',['wickr_identity',['../structwickr__identity.html',1,'wickr_identity'],['../group__wickr__identity.html',1,'(Global Namespace)']]],
-  ['wickr_5fidentity_5fchain',['wickr_identity_chain',['../structwickr__identity__chain.html',1,'']]],
-  ['wickr_5fidentity_5fchain_5fcopy',['wickr_identity_chain_copy',['../group__wickr__identity.html#gac9d708d340abcd0b54d87434b71d2d9e',1,'identity.h']]],
-  ['wickr_5fidentity_5fchain_5fcreate',['wickr_identity_chain_create',['../group__wickr__identity.html#ga79551385194ba6b044b676ce5c7400e6',1,'identity.h']]],
-  ['wickr_5fidentity_5fchain_5fdestroy',['wickr_identity_chain_destroy',['../group__wickr__identity.html#gadbddf6d49fddb2df103486970b3d564b',1,'identity.h']]],
-  ['wickr_5fidentity_5fchain_5fstatus',['wickr_identity_chain_status',['../group__wickr__identity.html#ga78aefa3835eb13ca75c0129583b5ebfa',1,'identity.h']]],
-  ['wickr_5fidentity_5fchain_5fvalidate',['wickr_identity_chain_validate',['../group__wickr__identity.html#ga12f5b33c4947b6a2ca3df3bdaacafa10',1,'identity.h']]],
-  ['wickr_5fidentity_5fcopy',['wickr_identity_copy',['../group__wickr__identity.html#ga14d1fb3ddd1dbd7797eb9b628eef99fe',1,'identity.h']]],
-  ['wickr_5fidentity_5fcreate',['wickr_identity_create',['../group__wickr__identity.html#ga75bfef275a250a32784231e8c042913b',1,'identity.h']]],
-  ['wickr_5fidentity_5fcreate_5ffrom_5fbuffer',['wickr_identity_create_from_buffer',['../group__wickr__identity.html#ga378e5b6ddabfe2c6ae7dbd5bdc43278e',1,'identity.h']]],
-  ['wickr_5fidentity_5fdestroy',['wickr_identity_destroy',['../group__wickr__identity.html#ga9e4a0f3736d52836c9a9fc61794c8ddd',1,'identity.h']]],
-  ['wickr_5fidentity_5fserialize',['wickr_identity_serialize',['../group__wickr__identity.html#ga79ca0713d0b90873e584379f26cbd051',1,'identity.h']]],
-  ['wickr_5fidentity_5fsign',['wickr_identity_sign',['../group__wickr__identity.html#ga32097764e57e70d85a2e1277146495c9',1,'identity.h']]],
-  ['wickr_5fkdf',['wickr_kdf',['../group__wickr__kdf.html',1,'']]],
-  ['wickr_5fkdf_5falgo',['wickr_kdf_algo',['../structwickr__kdf__algo.html',1,'']]],
-  ['wickr_5fkdf_5falgo_5fid',['wickr_kdf_algo_id',['../group__wickr__kdf.html#ga106149e3507106abb6b5be843a4e7c16',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta',['wickr_kdf_meta',['../structwickr__kdf__meta.html',1,'']]],
-  ['wickr_5fkdf_5fmeta_5fcopy',['wickr_kdf_meta_copy',['../group__wickr__kdf.html#ga6ed878108fd58ee420239689de505190',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta_5fcreate',['wickr_kdf_meta_create',['../group__wickr__kdf.html#gaeee99bcfbf94fd964158ac402e2deb48',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta_5fcreate_5fwith_5fbuffer',['wickr_kdf_meta_create_with_buffer',['../group__wickr__kdf.html#gaebd1caf94717c403c4afb5cca8f511b6',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta_5fdestroy',['wickr_kdf_meta_destroy',['../group__wickr__kdf.html#ga2bd458ae83d4d5bb03b9d0835baeefd7',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta_5fserialize',['wickr_kdf_meta_serialize',['../group__wickr__kdf.html#ga1bc501706b4a8a727dfd34f8a00e56da',1,'kdf.h']]],
-  ['wickr_5fkdf_5fmeta_5fsize_5fwith_5fbuffer',['wickr_kdf_meta_size_with_buffer',['../group__wickr__kdf.html#ga92020ae3e8ffb469c927d17f95a7dff1',1,'kdf.h']]],
-  ['wickr_5fkdf_5fresult',['wickr_kdf_result',['../structwickr__kdf__result.html',1,'']]],
-  ['wickr_5fkdf_5fresult_5fcopy',['wickr_kdf_result_copy',['../group__wickr__kdf.html#ga69c03c94bf446471ba74f76ee883f8d4',1,'kdf.h']]],
-  ['wickr_5fkdf_5fresult_5fcreate',['wickr_kdf_result_create',['../group__wickr__kdf.html#ga514c095b3b2f1ead49bd7fbf0b427407',1,'kdf.h']]],
-  ['wickr_5fkdf_5fresult_5fdestroy',['wickr_kdf_result_destroy',['../group__wickr__kdf.html#ga61ff0036a7257321ff0bbdc184d856a3',1,'kdf.h']]],
-  ['wickr_5fkey_5fexchange',['wickr_key_exchange',['../structwickr__key__exchange.html',1,'']]],
-  ['wickr_5fkey_5fexchange_5fcopy',['wickr_key_exchange_copy',['../group__wickr__protocol.html#ga2235a7c168e3806aa1a71b5a5c8dee48',1,'protocol.h']]],
-  ['wickr_5fkey_5fexchange_5fcreate',['wickr_key_exchange_create',['../group__wickr__protocol.html#gac2823b3430070b0db9ef1f40d3a804ea',1,'protocol.h']]],
-  ['wickr_5fkey_5fexchange_5fcreate_5fwith_5fdata',['wickr_key_exchange_create_with_data',['../group__wickr__protocol.html#ga9e69db5af6825e72c443dd5ef3605a5b',1,'protocol.h']]],
-  ['wickr_5fkey_5fexchange_5fcreate_5fwith_5fpacket_5fkey',['wickr_key_exchange_create_with_packet_key',['../group__wickr__protocol.html#ga26d4b847f11801354e4111d3ec1d1d13',1,'protocol.h']]],
-  ['wickr_5fkey_5fexchange_5fdestroy',['wickr_key_exchange_destroy',['../group__wickr__protocol.html#ga54b6481f8f4b61fb949bec17d47cd3de',1,'protocol.h']]],
-  ['wickr_5fkey_5fexchange_5fkdf_5fmatching_5fcipher',['wickr_key_exchange_kdf_matching_cipher',['../group__wickr__crypto__engine.html#ga6cdd263ddccd0b28a85ed61889805e60',1,'crypto_engine.h']]],
-  ['wickr_5fnode',['wickr_node',['../structwickr__node.html',1,'wickr_node'],['../group__wickr__node.html',1,'(Global Namespace)']]],
-  ['wickr_5fnode_5farray_5fcopy',['wickr_node_array_copy',['../group__wickr__node.html#gab577103b1752357b8575b064e22a4970',1,'node.h']]],
-  ['wickr_5fnode_5farray_5fdestroy',['wickr_node_array_destroy',['../group__wickr__node.html#gab3f40f122baca50bcc62af54eaa08fe9',1,'node.h']]],
-  ['wickr_5fnode_5farray_5ffetch_5fitem',['wickr_node_array_fetch_item',['../group__wickr__node.html#ga878a3015abc8563d512c0e5acf10df60',1,'node.h']]],
-  ['wickr_5fnode_5farray_5fnew',['wickr_node_array_new',['../group__wickr__node.html#ga504ebe707aa003032c4ac8044d73ee92',1,'node.h']]],
-  ['wickr_5fnode_5farray_5fset_5fitem',['wickr_node_array_set_item',['../group__wickr__node.html#gae511a75489982fc2e90271be63d9d753',1,'node.h']]],
-  ['wickr_5fnode_5fcopy',['wickr_node_copy',['../group__wickr__node.html#ga393c4fb42022fdc6f4336a62dbc7af91',1,'node.h']]],
-  ['wickr_5fnode_5fcreate',['wickr_node_create',['../group__wickr__node.html#ga293b1210648e2db5449be23fe0f3e644',1,'node.h']]],
-  ['wickr_5fnode_5fcreate_5ffrom_5fbuffer',['wickr_node_create_from_buffer',['../group__wickr__node.html#ga3906f6da9fbd61172d2d65466773aa83',1,'node.h']]],
-  ['wickr_5fnode_5fdestroy',['wickr_node_destroy',['../group__wickr__node.html#ga6acf07885403d9df6baad36a8d797f73',1,'node.h']]],
-  ['wickr_5fnode_5fidentity_5fgen',['wickr_node_identity_gen',['../group__wickr__identity.html#gae820c1efc090fef6a8d27dd700316023',1,'identity.h']]],
-  ['wickr_5fnode_5frotate_5fkeypair',['wickr_node_rotate_keypair',['../group__wickr__node.html#ga2ed650b947df7316cf0f915805ac5aa4',1,'node.h']]],
-  ['wickr_5fnode_5fserialize',['wickr_node_serialize',['../group__wickr__node.html#ga6012a4fe980b1a616369cbc4b0a95f74',1,'node.h']]],
-  ['wickr_5fnode_5fverify_5fsignature_5fchain',['wickr_node_verify_signature_chain',['../group__wickr__node.html#gac9e2d96e9109590124b2fd05a8826c1d',1,'node.h']]],
-  ['wickr_5fpacket',['wickr_packet',['../structwickr__packet.html',1,'']]],
-  ['wickr_5fpacket_5fcopy',['wickr_packet_copy',['../group__wickr__protocol.html#ga0c95ad86819bacf98c11115d748039db',1,'protocol.h']]],
-  ['wickr_5fpacket_5fcreate',['wickr_packet_create',['../group__wickr__protocol.html#gac952913ddaf848d2def181cd55b30883',1,'protocol.h']]],
-  ['wickr_5fpacket_5fcreate_5ffrom_5fbuffer',['wickr_packet_create_from_buffer',['../group__wickr__protocol.html#ga3427a40815c9f65a78516405d1b88072',1,'protocol.h']]],
-  ['wickr_5fpacket_5fcreate_5ffrom_5fcomponents',['wickr_packet_create_from_components',['../group__wickr__protocol.html#gaf34d8082df52a7ade2fea54eea65d342',1,'protocol.h']]],
-  ['wickr_5fpacket_5fdestroy',['wickr_packet_destroy',['../group__wickr__protocol.html#gab15d7556ad2806aa823ec2d3e3bf5a46',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader',['wickr_packet_header',['../structwickr__packet__header.html',1,'']]],
-  ['wickr_5fpacket_5fheader_5fcopy',['wickr_packet_header_copy',['../group__wickr__protocol.html#ga5ffb8763c9ebf2ad7322d507d117d7ba',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader_5fcreate',['wickr_packet_header_create',['../group__wickr__protocol.html#ga8ad519d1b5dc3960c426db9a556404e1',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader_5fcreate_5ffrom_5fcipher',['wickr_packet_header_create_from_cipher',['../group__wickr__protocol.html#ga53ff33471ac0c4c7e3feedfcce520639',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader_5fdestroy',['wickr_packet_header_destroy',['../group__wickr__protocol.html#gacb41cb9ad58871f01d8660e3ab115f4e',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader_5fencrypt',['wickr_packet_header_encrypt',['../group__wickr__protocol.html#ga3495440a494c33a38ce4154b400e9809',1,'protocol.h']]],
-  ['wickr_5fpacket_5fheader_5ffind',['wickr_packet_header_find',['../group__wickr__protocol.html#gac9fab2b713e40a607735bcce0befee96',1,'protocol.h']]],
-  ['wickr_5fpacket_5fmeta',['wickr_packet_meta',['../structwickr__packet__meta.html',1,'']]],
-  ['wickr_5fpacket_5fmeta_5fcopy',['wickr_packet_meta_copy',['../group__wickr__protocol.html#ga1bb5a0f2a49069679b174b3e6521fe9b',1,'protocol.h']]],
-  ['wickr_5fpacket_5fmeta_5fcreate',['wickr_packet_meta_create',['../group__wickr__protocol.html#ga146144d74da12c1fff766462fe7fa661',1,'protocol.h']]],
-  ['wickr_5fpacket_5fmeta_5fdestroy',['wickr_packet_meta_destroy',['../group__wickr__protocol.html#gad165ad670da761478cf5f8f91223c1d3',1,'protocol.h']]],
-  ['wickr_5fpacket_5fserialize',['wickr_packet_serialize',['../group__wickr__protocol.html#gaa983fe863635b69d69d6d86d01443ecf',1,'protocol.h']]],
-  ['wickr_5fparse_5fresult',['wickr_parse_result',['../structwickr__parse__result.html',1,'']]],
-  ['wickr_5fparse_5fresult_5fcopy',['wickr_parse_result_copy',['../group__wickr__protocol.html#ga1671066ddc19b38ed05d4460d5a96706',1,'protocol.h']]],
-  ['wickr_5fparse_5fresult_5fcreate_5ffailure',['wickr_parse_result_create_failure',['../group__wickr__protocol.html#gac2673bcfeae2f8d693e523809bb6bb30',1,'protocol.h']]],
-  ['wickr_5fparse_5fresult_5fcreate_5fsuccess',['wickr_parse_result_create_success',['../group__wickr__protocol.html#gae9d6837e2cea5b64993c124e671bd25f',1,'protocol.h']]],
-  ['wickr_5fparse_5fresult_5fdestroy',['wickr_parse_result_destroy',['../group__wickr__protocol.html#ga706d39419c8991f417f37f556cee8df9',1,'protocol.h']]],
-  ['wickr_5fparse_5fresult_5ffrom_5fpacket',['wickr_parse_result_from_packet',['../group__wickr__protocol.html#ga1654ef299f9189b27e5e58485f1ffcf5',1,'protocol.h']]],
-  ['wickr_5fpayload',['wickr_payload',['../structwickr__payload.html',1,'']]],
-  ['wickr_5fpayload_5fcopy',['wickr_payload_copy',['../group__wickr__protocol.html#ga96e1b0eabf307d112985946cbe2f6969',1,'protocol.h']]],
-  ['wickr_5fpayload_5fcreate',['wickr_payload_create',['../group__wickr__protocol.html#gad7d51d39a8477d87557536ff42399186',1,'protocol.h']]],
-  ['wickr_5fpayload_5fcreate_5ffrom_5fcipher',['wickr_payload_create_from_cipher',['../group__wickr__protocol.html#gaeeebf95ba580b92254a91009245028ee',1,'protocol.h']]],
-  ['wickr_5fpayload_5fdestroy',['wickr_payload_destroy',['../group__wickr__protocol.html#ga821c48aa748408e47ba2149e1628a487',1,'protocol.h']]],
-  ['wickr_5fpayload_5fencrypt',['wickr_payload_encrypt',['../group__wickr__protocol.html#ga7e4fd2ad4de22c2f7d6160400d2e15e7',1,'protocol.h']]],
-  ['wickr_5fperform_5fkdf',['wickr_perform_kdf',['../group__wickr__kdf.html#gabc745cc1349f64c0f25d6d12de5f64cd',1,'kdf.h']]],
-  ['wickr_5fperform_5fkdf_5fmeta',['wickr_perform_kdf_meta',['../group__wickr__kdf.html#ga69674856e11652a18663fc7f833ad974',1,'kdf.h']]],
-  ['wickr_20protocol',['Wickr Protocol',['../group__wickr__protocol.html',1,'']]],
-  ['wickr_5froot_5fkeys',['wickr_root_keys',['../structwickr__root__keys.html',1,'wickr_root_keys'],['../group__wickr__root__keys.html',1,'(Global Namespace)']]],
-  ['wickr_5froot_5fkeys_5fcopy',['wickr_root_keys_copy',['../group__wickr__root__keys.html#gadda1b8fd1b5cc1cf6b7b2b0e72912858',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fcreate',['wickr_root_keys_create',['../group__wickr__root__keys.html#gab6289774ce4bcbdc08584695f9b7c34c',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fcreate_5ffrom_5fbuffer',['wickr_root_keys_create_from_buffer',['../group__wickr__root__keys.html#ga5ab25c65aa25fbe96d978f8e06376653',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fdestroy',['wickr_root_keys_destroy',['../group__wickr__root__keys.html#gacabc33b7c56cbc84a5085d0a4330694f',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fexport',['wickr_root_keys_export',['../group__wickr__root__keys.html#ga6e3a04e7e1c65ada93450051eb4f4459',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fgenerate',['wickr_root_keys_generate',['../group__wickr__root__keys.html#gac6d33868c25a85e5da5cff92a4eab189',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5flocalize',['wickr_root_keys_localize',['../group__wickr__root__keys.html#gaa34dd6c85c667b8b7fb025f56c7bcc46',1,'root_keys.h']]],
-  ['wickr_5froot_5fkeys_5fserialize',['wickr_root_keys_serialize',['../group__wickr__root__keys.html#ga6875c9f677c4ec127276714976a623a5',1,'root_keys.h']]],
-  ['wickr_5fstorage_5fkeys',['wickr_storage_keys',['../structwickr__storage__keys.html',1,'wickr_storage_keys'],['../group__wickr__storage__keys.html',1,'(Global Namespace)']]],
-  ['wickr_5fstorage_5fkeys_5fcopy',['wickr_storage_keys_copy',['../group__wickr__storage__keys.html#gad6839a21bca1c28775d51deb6b7c3835',1,'storage.h']]],
-  ['wickr_5fstorage_5fkeys_5fcreate',['wickr_storage_keys_create',['../group__wickr__storage__keys.html#gad03ddd5ea10633ca35097e5f4707d5cb',1,'storage.h']]],
-  ['wickr_5fstorage_5fkeys_5fcreate_5ffrom_5fbuffer',['wickr_storage_keys_create_from_buffer',['../group__wickr__storage__keys.html#gae5631d93182e635fb7f48269eb3f1723',1,'storage.h']]],
-  ['wickr_5fstorage_5fkeys_5fdestroy',['wickr_storage_keys_destroy',['../group__wickr__storage__keys.html#ga2c1494802577d025d70bf553c43da344',1,'storage.h']]],
-  ['wickr_5fstorage_5fkeys_5fserialize',['wickr_storage_keys_serialize',['../group__wickr__storage__keys.html#ga10a0d6e13442682b2fecff2ed753c2f6',1,'storage.h']]],
-  ['wickr_5fstream',['wickr_stream',['../group__wickr__stream.html',1,'']]],
-  ['wickr_5fstream_5fctx',['wickr_stream_ctx',['../structwickr__stream__ctx.html',1,'']]],
-  ['wickr_5fstream_5fctx_5fcopy',['wickr_stream_ctx_copy',['../group__wickr__stream.html#ga707b685e35eb437c33f5ec32c6760005',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fctx_5fcreate',['wickr_stream_ctx_create',['../group__wickr__stream.html#ga0b5c513a3e0958dc7fd4f8455206e32e',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fctx_5fdecode',['wickr_stream_ctx_decode',['../group__wickr__stream.html#gaa1eb323694f2c1652b51127b0b89138c',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fctx_5fdestroy',['wickr_stream_ctx_destroy',['../group__wickr__stream.html#ga05260bfce1f50352cad32e0354c28a1d',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fctx_5fencode',['wickr_stream_ctx_encode',['../group__wickr__stream.html#ga621a3ec801ab996b4fab1fe905c450bd',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fiv',['wickr_stream_iv',['../structwickr__stream__iv.html',1,'wickr_stream_iv'],['../group__wickr__stream__iv.html',1,'(Global Namespace)']]],
-  ['wickr_5fstream_5fiv_5fcopy',['wickr_stream_iv_copy',['../group__wickr__stream__iv.html#gaec8a5b5d70a86997c8158d047c7eed34',1,'stream_iv.h']]],
-  ['wickr_5fstream_5fiv_5fcreate',['wickr_stream_iv_create',['../group__wickr__stream__iv.html#gaaad5726a228b7866b0cb392c131f95be',1,'stream_iv.h']]],
-  ['wickr_5fstream_5fiv_5fdestroy',['wickr_stream_iv_destroy',['../group__wickr__stream__iv.html#gad6ac87d891822a866a0d7cad48f79e27',1,'stream_iv.h']]],
-  ['wickr_5fstream_5fiv_5fgenerate',['wickr_stream_iv_generate',['../group__wickr__stream__iv.html#gaddcfdc9904ec0ffd32ce8f988aefc484',1,'stream_iv.h']]],
-  ['wickr_5fstream_5fkey',['wickr_stream_key',['../structwickr__stream__key.html',1,'']]],
-  ['wickr_5fstream_5fkey_5fcopy',['wickr_stream_key_copy',['../group__wickr__stream.html#ga0201dd73adf773834d7dbb4bbc307fae',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fkey_5fcreate',['wickr_stream_key_create',['../group__wickr__stream.html#ga54449667a25d9e64cd04dae58dd76418',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fkey_5fcreate_5ffrom_5fbuffer',['wickr_stream_key_create_from_buffer',['../group__wickr__stream.html#gadf1565329fed927531e1a14334ed23e4',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fkey_5fcreate_5frand',['wickr_stream_key_create_rand',['../group__wickr__stream.html#ga9e95dd486dbe2f3db647c65042601b34',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fkey_5fdestroy',['wickr_stream_key_destroy',['../group__wickr__stream.html#ga3bb3dec760c5b7f3a1bd2d0bb13bcc79',1,'stream_cipher.h']]],
-  ['wickr_5fstream_5fkey_5fserialize',['wickr_stream_key_serialize',['../group__wickr__stream.html#gae1540c98a88e2f62e96fb2ee209e2290',1,'stream_cipher.h']]],
-  ['wickr_5ftransport_5fcallbacks',['wickr_transport_callbacks',['../structwickr__transport__callbacks.html',1,'']]],
-  ['wickr_5ftransport_5fctx',['wickr_transport_ctx',['../structwickr__transport__ctx.html',1,'wickr_transport_ctx'],['../group__wickr__transport__ctx.html',1,'(Global Namespace)']]],
-  ['wickr_5ftransport_5fctx_5fcopy',['wickr_transport_ctx_copy',['../group__wickr__transport__ctx.html#gaa49b2cf44bcb02d656598a17e5ae1c52',1,'transport_ctx.h']]],
-  ['wickr_5ftransport_5fctx_5fcreate',['wickr_transport_ctx_create',['../group__wickr__transport__ctx.html#ga4731758cf20b71d4469c74904bf662c0',1,'transport_ctx.h']]],
-  ['wickr_5ftransport_5fctx_5fdestroy',['wickr_transport_ctx_destroy',['../group__wickr__transport__ctx.html#gabe547b01f4f02efea3f518d7445a71ee',1,'transport_ctx.h']]],
-  ['wickr_5ftransport_5fctx_5fprocess_5frx_5fbuffer',['wickr_transport_ctx_process_rx_buffer',['../group__wickr__transport__ctx.html#gad982fd3463a81a667592e18d7bf90822',1,'transport_ctx.h']]],
-  ['wickr_5ftransport_5fctx_5fprocess_5ftx_5fbuffer',['wickr_transport_ctx_process_tx_buffer',['../group__wickr__transport__ctx.html#gad9d2f86832cfb9c1948ba679fabbedbc',1,'transport_ctx.h']]],
-  ['wickr_5ftransport_5fctx_5fstart',['wickr_transport_ctx_start',['../group__wickr__transport__ctx.html#gaee74761e31d5f0a6fc97e2ccf5ac1e29',1,'transport_ctx.h']]]
+  ['version',['version',['../structwickr__packet.html#a35c17c926c267a603e810b88961c2bf3',1,'wickr_packet']]]
 ];
