+++ docs/html/search/searchdata.js
@@ -1,9 +1,9 @@
 var indexSectionsWithContent =
 {
-  0: "abcdefghiklmnoprstvw",
+  0: "abcdefghiklmnoprstuvw",
   1: "w",
   2: "gow",
-  3: "abcdeghiklmnoprstvw",
+  3: "abcdeghiklmnoprstuvw",
   4: "w",
   5: "dfmopw"
 };
