+++ docs/html/search/variables_12.js
@@ -1,20 +1,4 @@
 var searchData=
 [
-  ['wickr_5fcrypto_5fengine_5fcipher_5fdecrypt',['wickr_crypto_engine_cipher_decrypt',['../group__wickr__crypto__engine.html#ga459ef821cc40edb5e3c9ce1d571ddc00',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcipher_5fencrypt',['wickr_crypto_engine_cipher_encrypt',['../group__wickr__crypto__engine.html#gaf7be0d0ca0d9e7a0c5c3b1ec0fd0918e',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcipher_5fkey_5frandom',['wickr_crypto_engine_cipher_key_random',['../group__wickr__crypto__engine.html#gabc120effc5248a491b18138f9b74b526',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fcrypto_5frandom',['wickr_crypto_engine_crypto_random',['../group__wickr__crypto__engine.html#gae12855853c05caa473e748a62611d307',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdecrypt_5ffile',['wickr_crypto_engine_decrypt_file',['../group__wickr__crypto__engine.html#ga361b843efc839e30f90ce24c15b1dd27',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdigest',['wickr_crypto_engine_digest',['../group__wickr__crypto__engine.html#ga42dfc5b4df166616d962ee4c792ceaf4',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fdigest_5ffile',['wickr_crypto_engine_digest_file',['../group__wickr__crypto__engine.html#ga298d2dea6722151c11e04458d3924823',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fkey_5fimport',['wickr_crypto_engine_ec_key_import',['../group__wickr__crypto__engine.html#ga87a069bbd3cdf509946e4ca0ca008f8f',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5frand_5fkey',['wickr_crypto_engine_ec_rand_key',['../group__wickr__crypto__engine.html#gad16b229b7d16cf00d3983c277ee3b88b',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fsign',['wickr_crypto_engine_ec_sign',['../group__wickr__crypto__engine.html#gac5a36d2a53e06a75a2b818e6a5cc96a1',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fec_5fverify',['wickr_crypto_engine_ec_verify',['../group__wickr__crypto__engine.html#gae541acc6433f6f5ffe84382ccd5e4f97',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fecdh_5fgen_5fkey',['wickr_crypto_engine_ecdh_gen_key',['../group__wickr__crypto__engine.html#ga94bdc8d3b88ac04b9defe663afa48e78',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fencrypt_5ffile',['wickr_crypto_engine_encrypt_file',['../group__wickr__crypto__engine.html#gaa77b1bc373d92d9db9d801bcc4a4d29c',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fhmac_5fcreate',['wickr_crypto_engine_hmac_create',['../group__wickr__crypto__engine.html#ga782e0e8786ecf8a55c8ae2aea2f178aa',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fengine_5fhmac_5fverify',['wickr_crypto_engine_hmac_verify',['../group__wickr__crypto__engine.html#ga00c52a816403192af5b21d952265d0b6',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fkdf_5fgen',['wickr_crypto_kdf_gen',['../group__wickr__crypto__engine.html#ga5ab29968c993423df83c9f1a3d7e685d',1,'wickr_crypto_engine']]],
-  ['wickr_5fcrypto_5fkdf_5fmeta',['wickr_crypto_kdf_meta',['../group__wickr__crypto__engine.html#ga640cbc86ca36c4deb2c197de89dce7d1',1,'wickr_crypto_engine']]]
+  ['version',['version',['../structwickr__packet.html#a35c17c926c267a603e810b88961c2bf3',1,'wickr_packet']]]
 ];
