+++ docs/latex/modules.tex
@@ -22,6 +22,7 @@ Here is a list of all modules\+:\begin{DoxyCompactList}
 \item \contentsline{section}{Wickr\+\_\+storage\+\_\+keys}{\pageref{group__wickr__storage__keys}}{}
 \item \contentsline{section}{wickr\+\_\+stream}{\pageref{group__wickr__stream}}{}
 \item \contentsline{section}{Wickr\+\_\+stream\+\_\+iv}{\pageref{group__wickr__stream__iv}}{}
+\item \contentsline{section}{wickr\+\_\+stream\+\_\+key}{\pageref{group__wickr__stream__key}}{}
 \item \contentsline{section}{Wickr\+\_\+transport\+\_\+ctx}{\pageref{group__wickr__transport__ctx}}{}
 \item \contentsline{section}{Data Conversion Utilities}{\pageref{group__data__conversion__utilities}}{}
 \item \contentsline{section}{Wickr\+\_\+ctx}{\pageref{group__wickr__ctx}}{}
