+++ docs/latex/refman.tex
@@ -165,6 +165,7 @@
 \include{group__wickr__storage__keys}
 \include{group__wickr__stream}
 \include{group__wickr__stream__iv}
+\include{group__wickr__stream__key}
 \include{group__wickr__transport__ctx}
 \include{group__data__conversion__utilities}
 \include{group__wickr__ctx}
