+++ src/crypto/proto/stream.proto
@@ -33,13 +33,9 @@ message Handshake {
     }
 }
 
-message StreamInfo {
-    optional StreamKey key = 1;
-    optional bytes user_data = 2;
-}
-
 message StreamKey {
     optional bytes cipher_key = 1;
     optional bytes evolution_key = 2;
     optional uint32 packets_per_evo = 3;
+    optional bytes user_data = 4;
 }
