+++ src/crypto/stream_ctx.h
@@ -19,44 +19,18 @@
  * ACCESSING AND/OR USING THE CODE ON LICENSEE’S SYSTEM.
  */
 
-#ifndef stream_cipher_h
-#define stream_cipher_h
+#ifndef stream_ctx_h
+#define stream_ctx_h
 
 #include "crypto_engine.h"
 #include "stream_iv.h"
-
-#define PACKET_PER_EVO_MIN 64
-#define PACKET_PER_EVO_DEFAULT 512
-#define PACKET_PER_EVO_MAX 32768
-
-typedef enum { STREAM_DIRECTION_ENCODE, STREAM_DIRECTION_DECODE } wickr_stream_direction;
+#include "stream_key.h"
 
 /**
  @addtogroup wickr_stream wickr_stream
  */
 
-/**
- @ingroup wickr_stream
- @struct wickr_stream_key
- 
- @brief A data structure representing the stream encoding / decoding key material
- A stream key holds information about the key material used for cipher operations as well as it's next evolution
- key and the number of packets this key should be used to encode or decode before evolution takes place
- 
- @var wickr_stream_key::cipher_key
- key used to encrypt or decrypt packets when the key is used for cipher operations
- @var wickr_stream_key::evolution_key
- data to be used to help evolove the key when 'cipher_key' is used 'packets_per_evolution' times
- @var wickr_stream_key::packets_per_evolution
- number of packets this key should be used before it is evoloved using 'evolution_key'
- */
-struct wickr_stream_key {
-    wickr_cipher_key_t *cipher_key;
-    wickr_buffer_t *evolution_key;
-    uint32_t packets_per_evolution;
-};
-
-typedef struct wickr_stream_key wickr_stream_key_t;
+typedef enum { STREAM_DIRECTION_ENCODE, STREAM_DIRECTION_DECODE } wickr_stream_direction;
 
 /**
  @ingroup wickr_stream
@@ -75,10 +49,10 @@ typedef struct wickr_stream_key wickr_stream_key_t;
  @var wickr_stream_ctx::engine
  crypto engine to be used for cipher operations, as well as key evolution using HMAC
  @var wickr_stream_ctx::key
- the current stream key that the stream context is using for encryption or decryption depending on directoin
+ the current stream key that the stream context is using for encryption or decryption depending on direction
  the stream key will change over the course of packet encoding due to it's evolutions
  @var wickr_stream_ctx::iv_factory
- the iv generator that this stream is using to create iv's
+ the iv generator that this stream is using to create IVs
  @var wickr_stream_ctx::last_seq
  the most recent sequence number that successfully encrypted or decrypted a packet
  @var wickr_stream_ctx::direction
@@ -93,71 +67,6 @@ struct wickr_stream_ctx {
 };
 
 typedef struct wickr_stream_ctx wickr_stream_ctx_t;
-
-/**
- 
- @ingroup wickr_stream
- 
- Create a stream key from components
-
- @param cipher_key see documentation of 'wickr_stream_key' structure
- @param evolution_key see documentation of 'wickr_stream_key' structure
- @param packets_per_evolution see documentation of 'wickr_stream_key' structure
- @return a newly allocated stream key owning the properties passed in
- */
-wickr_stream_key_t *wickr_stream_key_create(wickr_cipher_key_t *cipher_key, wickr_buffer_t *evolution_key, uint32_t packets_per_evolution);
-
-/**
- @ingroup wickr_stream
- 
- Generate a random stream key
-
- @param engine the crypto engine to use for secure random cipher key generation
- @param cipher the cipher to use for generation of the internal 'cipher_key' property
- @param packets_per_evolution the number of times this key should be used before it evoloves
- @return a newly allocated stream key
- */
-wickr_stream_key_t *wickr_stream_key_create_rand(const wickr_crypto_engine_t engine, wickr_cipher_t cipher, uint32_t packets_per_evolution);
-
-/**
- @ingroup wickr_stream
- 
- Copy a stream key
-
- @param stream_key the stream key to copy
- @return a newly allocated stream key holding a deep copy of properties from 'stream_key'
- */
-wickr_stream_key_t *wickr_stream_key_copy(const wickr_stream_key_t *stream_key);
-
-/**
- @ingroup wickr_stream
- 
- Serialize a stream key
-
- @param key the key to serialize
- @return a serialized protocol buffer object representing the properties of 'key'
- */
-wickr_buffer_t *wickr_stream_key_serialize(const wickr_stream_key_t *key);
-
-/**
- @ingroup wickr_stream
- 
- Create a stream key from a serialized buffer
-
- @param buffer the buffer to parse into a stream key
- @return a newly allocated stream key represented by 'buffer' or NULL if parsing buffer fails
- */
-wickr_stream_key_t *wickr_stream_key_create_from_buffer(const wickr_buffer_t *buffer);
-
-/**
- 
- @ingroup wickr_stream
- 
- Destroy a stream key
- 
- @param stream_key a pointer to the stream key to destroy. All properties of '*stream_key' will also be destroyed
- */
-void wickr_stream_key_destroy(wickr_stream_key_t **stream_key);
 
 /**
  @ingroup wickr_stream
@@ -216,4 +125,4 @@ wickr_buffer_t *wickr_stream_ctx_decode(wickr_stream_ctx_t *ctx, const wickr_cip
  */
 void wickr_stream_ctx_destroy(wickr_stream_ctx_t **ctx);
 
-#endif /* stream_cipher_h */
+#endif /* stream_ctx_h */
