+++ src/crypto/transport_ctx.c
@@ -10,7 +10,7 @@
 #include "memory.h"
 #include "stream.pb-c.h"
 #include "protobuf_util.h"
-#include "stream_cipher.h"
+#include "stream_ctx.h"
 #include "protocol.h"
 #include "transport_priv.h"
 #include "node_priv.h"
